Fix issue with eliminatePhi.
We used to eliminate Phi nodes by inserting a copy at the end of the
incoming block.  This approach breaks when the CFG is modified.  We now
simply insert the copy after the incoming variable has been written to.
Since this is SSA, we don't have to worry about updates.